Update README.md
The goal is to make the config-based ("quickstart") steps prominent and enable the docs to link directly to them. Therefore, I put those steps in the main README file, and I propose removing the https://github.com/wavefrontHQ/wavefront-jersey-sdk-java/blob/master/docs/basic-mode.md file. (Links I've added to the advanced-mode.md file in a separate pull request assume that the quickstart steps are now on the main README page.)