Update README.md (#40)
* Update README.md

Incorporated `WavefrontJaxrsClientFilter` into Quickstart steps.

* update custom.md

Added steps for creating a `WavefrontJaxrsClientFilter` to the custom steps.

* Update dropwizard.md

Copy edits. Updated link for creating a WavefrontJerseyFilter. Now there are 2 links (quickstart and custom steps). Link destination in this file assumes the changes made in this PR to README.md and custom.md.

* Update springboot.md

Copy edits. Updated link for creating a WavefrontJerseyFilter. Now there are 2 links (quickstart and custom steps). Link destination in this file assumes the changes made in this PR to README.md and custom.md.